+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Bugs:
 * #111: Generated urls for non-default ports included the port twice.
 * #28: Permission denied page.
 * #53: User selection on server edit page.
+* #115: Warning on server page when notifications are disabled.
 * Composer added for dependencies.
 
 v3.0.1 (released April 12, 2014)